requirements.txt
2captcha-python==1.2.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
dnspython==2.4.1
exceptiongroup==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.10
lxml==4.9.1
outcome==1.2.0
packaging==23.1
pyasn1==0.5.0
pycparser==2.21
pymongo==4.4.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.4.3
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.65.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==4.0.0
wsproto==1.2.0
zstandard==0.21.0
